requirements.txt
typing  # Version dependent on Python version.
absl-py>=1.0.0
numpy>=1.21.5,<2.0.0
scipy>=1.15.2

# Distributed systems libraries.
# NOTE: Requires tensorflow~=2.8.0 to avoid proto issues.
dm-launchpad[tensorflow]
dm-reverb[tensorflow]
ml-collections>=0.1.1
